174: extract function to create issue link into separate file
